Fix loading state bug.
Now when we're waiting for initial login state to be loaded before
showing messages about a user not being logged in. Previously it said
whenever user was null show user not logged in, but this happened
briefly on page load so this was not correct. Now we have a dedicated
loading state.